Move ean_product.dart into 'model' folder
