+++ bin/model/ean_product.dart
